refactor(voice): split oversized files into modules
VoiceSection.tsx (1604→791 lines):
- Extract PersonaEditModal, AddVoiceProviderModal, LocalProviderInstructions
  into new VoiceModals.tsx (847 lines)
- Shared interfaces and signals exported from VoiceModals

voice.rs (1551→3 files):
- voice/mod.rs (491 lines): config loading, key migration, helpers
- voice/tts_service.rs (602 lines): LiveTtsService + TtsService impl
- voice/stt_service.rs (545 lines): LiveSttService + SttService impl
- All public API paths unchanged via re-exports

All files under the 1500-line CI limit.
